backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
aiomysql
redis
pydantic
